tests(auth): include session type definition in test run
This fixes the failed test run for libs/data-access/api/src/lib/auth.spec.ts.
When the session type is not defined, the default user object is used. This object does not have a role attribute. By including the index.d.ts in the test, the correct type ius used.