fix(quiz-editor): Markdown editor is now opened directly when adding something to a Categorie (#342)
At first after a adding a Categorie in the Learncontrol editor a popup
was opened that you had to click on again to get to the markdown editor.
Now if you press add or the pencil icon you will immediatly get
redirected to the markdown-editor, where you can make your changes.