feat(layout): searchbar improvements (#327)
Refactored the Searchbar to a combobox and added Arrowkey-navigation as well as onFocus and onBlur methods.
The searchsection method was removed -> icons are now displayed at the searchresults, hinting which type of result it is.
The Question is, if a filter option for the searchsection(author,courses,learningunits) is needed.
Learningunit search still points at the wrong slug, because i have to wait for a change of the learningunit structure, so that they are searchable independent of courses.
Also did a quick fix on the placeholder of the Course-editor mentioned in Issue #84.