rf(achievements): add reusable option to convert achievements from DB to AchievementWithProgress type
This should make a refactor of the data flow in achievements more easy in the future.
Currently the profile loads necessary data via SSP. But the streak dialog loads all achievements every time on its own via trpc. In the future it is advised to change this behaviour so the profile SSP loads the achievements, converts them in the right format with the new convert method and then gives them to the streak dialog