fix(diary): Fixes wrong computation of total duration (#389)
This contains 2 fixes and 1 improvement:
- Fix: Correct computation of total duration: Computes the duration per
resource and per change. So if Resource A, Resource B, and again
Resource A are learned, two durations for A are computed and sumed up to
a total duration. This is done to:
  - Not consider breaks between Resource changes
  - Not computing the duration of Resource A twice
- Fix: Checks if endDate of a diary Page is later than 6 hours after
start (was requirement by Psychology). If yes the endDate is adjusted to
avoid consideration of events of shorter lesson visists, for which no
DiaryPage is created (less than 1 minute).
- Improvement: Metric moved to `analysis` to facilitate re-use.

Fixes #386